+++ my_strncmp.c
@@ -27,12 +27,12 @@
     thereof) is found, respectively, to be less than, to match, or be
     greater than s2.
 */
-int my_strncmp(const char *s1, const char *s2, int n)
+int my_strncmp(const char *s1, const char *s2, size_t n)
 {
     if (s1 == NULL || s2 == NULL)
         return 0;
 
-    int i;
+    size_t i;
     for (i = 0; i < n && s1[i] && s2[i]; i++)
         if (s1[i] != s2[i])
             return s1[i] - s2[i];
